requirements.txt
dark_matter==4.0.58
pysam==0.21.0
